+++ Reach-Coder/Changes.md
@@ -1,5 +1,11 @@
 # Reach Coder - List of changes
 
+## v1.0.2 (in development)
+### New features
+- Backups are now taken when files are merged together
+### Other changes
+- Generated scala files are now placed in a subdirectory called `src`
+
 ## v1.0.1 - 22.01.2024
 Supports **Reach v1.2**
 ### Bugfixes